Rename asap-genetics to asap-spectrum (#1264)
* Rename all file that contain 'genetics'

* replace all occurences of genetics to spectrum

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------
