+++ .github/workflows/macos-ci.yaml
@@ -84,7 +84,7 @@ jobs:
         python -m pip install -e ./asapdiscovery-ml --no-deps
         python -m pip install -e ./asapdiscovery-modeling --no-deps
         python -m pip install -e ./asapdiscovery-simulation --no-deps
-        python -m pip install -e ./asapdiscovery-genetics --no-deps
+        python -m pip install -e ./asapdiscovery-spectrum --no-deps
         micromamba list
 
     - name: Run tests
@@ -103,7 +103,7 @@ jobs:
                        --cov=asapdiscovery-ml \
                        --cov=asapdiscovery-modeling \
                        --cov=asapdiscovery-simulation \
-                        --cov=asapdiscovery-genetics \
+                        --cov=asapdiscovery-spectrum \
                        asapdiscovery-alchemy/asapdiscovery/alchemy/tests \
                        asapdiscovery-data/asapdiscovery/data/tests \
                        asapdiscovery-dataviz/asapdiscovery/dataviz/tests \
@@ -111,7 +111,7 @@ jobs:
                        asapdiscovery-ml/asapdiscovery/ml/tests \
                        asapdiscovery-modeling/asapdiscovery/modeling/tests \
                        asapdiscovery-simulation/asapdiscovery/simulation/tests \
-                       asapdiscovery-genetics/asapdiscovery/genetics/tests
+                       asapdiscovery-spectrum/asapdiscovery/spectrum/tests
 
     - name: Upload Code Coverage to Codecov
       uses: codecov/codecov-action@v3
