+++ .github/workflows/simulation.yaml
@@ -83,7 +83,7 @@ jobs:
         python -m pip install -e ./asapdiscovery-ml --no-deps
         python -m pip install -e ./asapdiscovery-modeling --no-deps
         python -m pip install -e ./asapdiscovery-simulation --no-deps
-        python -m pip install -e ./asapdiscovery-genetics --no-deps
+        python -m pip install -e ./asapdiscovery-spectrum --no-deps
         python -m pip install -e ./asapdiscovery-workflows --no-deps
         micromamba list
