+++ asapdiscovery-docking/asapdiscovery/docking/scorer.py
@@ -23,9 +23,9 @@ from asapdiscovery.data.util.dask_utils import (
 )
 from asapdiscovery.docking.docking import DockingResult
 from asapdiscovery.docking.docking_data_validation import DockingResultCols
-from asapdiscovery.genetics.fitness import target_has_fitness_data
 from asapdiscovery.ml.inference import InferenceBase, get_inference_cls_from_model_type
 from asapdiscovery.ml.models import MLModelSpecBase
+from asapdiscovery.spectrum.fitness import target_has_fitness_data
 from mtenn.config import ModelType
 from multimethod import multimethod
 from pydantic import BaseModel, Field, validator
