--- asapdiscovery-genetics/README.md
@@ -1,1 +1,0 @@
-# asapdiscovery-genetics
