+++ asapdiscovery-spectrum/LICENSE
