+++ asapdiscovery-spectrum/MANIFEST.in
