+++ asapdiscovery-spectrum/asapdiscovery/spectrum/__init__.py
@@ -2,4 +2,4 @@
 
 from importlib.metadata import version
 
-__version__ = version("asapdiscovery-genetics")
+__version__ = version("asapdiscovery-spectrum")
