+++ asapdiscovery-spectrum/pyproject.toml
@@ -3,7 +3,7 @@ requires = ["setuptools>=61.0", "versioningit"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "asapdiscovery-genetics"
+name = "asapdiscovery-spectrum"
 #version = "0.0.1"
 dynamic = ["version"]
 authors = [
@@ -25,7 +25,7 @@ dependencies = []
 "Bug Tracker" = "https://github.com/choderalab/asapdiscovery/issues"
 
 [project.scripts]
-asap-genetics = "asapdiscovery.genetics.cli:genetics"
+asap-spectrum = "asapdiscovery.spectrum.cli:spectrum"
 
 [tool.setuptools.packages.find]
 where = ["."]
