+++ asapdiscovery-workflows/asapdiscovery/workflows/docking_workflows/large_scale_docking.py
@@ -37,9 +37,9 @@ from asapdiscovery.docking.scorer import (
     MetaScorer,
     MLModelScorer,
 )
-from asapdiscovery.genetics.fitness import target_has_fitness_data
 from asapdiscovery.ml.models import ASAPMLModelRegistry
 from asapdiscovery.modeling.protein_prep import ProteinPrepper
+from asapdiscovery.spectrum.fitness import target_has_fitness_data
 from asapdiscovery.workflows.docking_workflows.workflows import (
     PosteraDockingWorkflowInputs,
 )
