+++ docs/modules/spectrum.md
@@ -1,2 +1,2 @@
-genetics
+spectrum
 ========
